[YDBTest#501] Fix 4-byte to 8-byte conversions missed out in wcs_verify.c/wcs_recover.c (in GT.M V7.0-000)
Background
----------
* ea9950a1 introduced a scheme to test huge sparse database files using a debug-only env var.
  Setting that env var to random values in the YDBTest framework, we found a few tests failed
  the following assert for certain settings of the env var.

  ```c
  %YDB-F-ASSERT, Assert failed in sr_port/wcs_recover.c line 285 for expression (FALSE)
  ```

* Below is a simple example that reproduces the problem using the debug-only env var.

  ```sh
  export ydb_test_4g_db_blks=16777215
  export ydb_gbldir=yottadb.gld
  rm -f yottadb.gld yottadb.dat
  echo "change -segment DEFAULT -file=yottadb.dat" | $ydb_dist/yottadb -run GDE
  $ydb_dist/mupip create
  $ydb_dist/yottadb -run %XCMD 'for i=1:1:10000 set ^x(i)=$j(i,200)'
  echo "dump -bl=1FFFFFE00\ncache -verify\n" | $ydb_dist/dse
  ```

* The last step in the above example resulted in the following output when run with a debug-build.

  ```c
  %YDB-F-ASSERT, Assert failed in sr_port/wcs_verify.c line 549 for expression (expect_damage)
  %YDB-F-ASSERT, Assert failed in sr_port/wcs_recover.c line 285 for expression (FALSE)
  ```

Issue
-----
* The above example uncovered places in the code where an 8-byte block number was typecast to a
  4-byte `int`.

* If the 8-byte block number happened to be in the range `2**31` to `2**32`, then this value when
  type cast to a 4-byte integer would end up being negative and would fail comparisons like
  `((int)(cr->blk) < 0)` when actually the value is not a negative number.

* This is a regression caused after merging GT.M V7.0-000. These places got missed out in the
  original GT.M update.

Fix
---
* All places in the code that used a `blk` field and `int` in the same line were searched and ensured
  no type cast of an 8-byte block number is done to `int`. If any were present, they were fixed in this
  commit by removing the `(int)` type cast. This is because `block_id` is now an 8-byte signed quantity
  (whereas before GT.M V7.0-000, it was an unsigned 4-byte quantity) and therefore using it in checks
  for negative numbers can be done directly without any type cast or signed cast.

  That is, one can just say `(cr->blk < 0)` (instead of `((int)(cr->blk) < 0)`) to check if the block
  number is less than 0. This will return true if the block has the high order bit set to 1 (i.e. the
  63rd bit in the 64-bit number) instead of returning true if the block has the 31st bit set to 1.

+++ sr_port/wcs_recover.c
@@ -278,7 +278,7 @@ void wcs_recover(gd_region *reg)
 	for (cr = cr_lo, total_rip_wait = 0; cr < cr_hi; cr++, buffptr += blk_size)
 	{
 		cr->buffaddr = GDS_ANY_ABS2REL(csa, buffptr);	/* reset it to what it should be just to be safe */
-		if (((int)(cr->blk) != CR_BLKEMPTY) && (((int)(cr->blk) < 0) || ((int)(cr->blk) >= csd->trans_hist.total_blks)))
+		if ((cr->blk != CR_BLKEMPTY) && ((cr->blk < 0) || (cr->blk >= csd->trans_hist.total_blks)))
 		{	/* bad block number. discard buffer for now. actually can do better by looking at cr->bt_index... */
 			cr->cycle++;	/* increment cycle whenever blk number changes (tp_hist depends on this) */
 			cr->blk = CR_BLKEMPTY;
